+++ src/state/include/actor/states/actor_state.h
@@ -39,7 +39,7 @@ public:
 	 */
 	virtual std::unique_ptr<ActorState> Update(
 		Actor * actor,
-		int64_t delta_time
+		float delta_time
 	) = 0;
 	/**
 	 * Called right before the Actor switches to another state
